refactored error message
