+++ src/vis/heatmap/interfaces.ts
@@ -1,4 +1,4 @@
-import { ColumnInfo, EAggregateTypes, ENumericalColorScaleType, ESortTypes, ESupportedPlotlyVis } from '../interfaces';
+import { ColumnInfo, EAggregateTypes, ENumericalColorScaleType, ESupportedPlotlyVis } from '../interfaces';
 
 export interface IHeatmapConfig {
   type: ESupportedPlotlyVis.HEATMAP;
@@ -8,4 +8,12 @@ export interface IHeatmapConfig {
   sortedBy: ESortTypes;
   aggregateType: EAggregateTypes;
   aggregateColumn: ColumnInfo | null;
+}
+
+export enum ESortTypes {
+  NONE = 'NONE',
+  CAT_ASC = 'CAT_ASC',
+  CAT_DESC = 'CAT_DESC',
+  COUNT_ASC = 'COUNT_ASC',
+  COUNT_DESC = 'COUNT_DESC',
 }
