+++ src/vis/raincloud/RaincloudGrid.tsx
@@ -1,10 +1,11 @@
 import { SimpleGrid } from '@mantine/core';
 import React from 'react';
 import { VisColumn } from '../interfaces';
-import { getRaincloudData } from './utils';
 import { IRaincloudConfig } from './interfaces';
+import { getRaincloudData } from './utils';
 
 import { useAsync } from '../../hooks/useAsync';
+import { InvalidCols } from '../general/InvalidCols';
 import { Raincloud } from './Raincloud';
 
 export function RaincloudGrid({
@@ -22,10 +23,13 @@ export function RaincloudGrid({
 
   return (
     <SimpleGrid cols={Math.ceil(Math.sqrt(data?.numColVals.length))} style={{ width: '100%', height: '100%' }}>
-      {data &&
+      {data && config.numColumnsSelected.length >= 1 ? (
         data.numColVals.map((numCol) => {
           return <Raincloud key={numCol.info.id} column={numCol} config={config} selectionCallback={selectionCallback} selected={selected} />;
-        })}
+        })
+      ) : (
+        <InvalidCols headerMessage="Invalid settings" bodyMessage="To create a raincloud chart, select at least 2 numerical columns." />
+      )}
     </SimpleGrid>
   );
 }
