+++ src/vis/raincloud/RaincloudRainSelect.tsx
@@ -1,7 +1,6 @@
 import { Select } from '@mantine/core';
 import * as React from 'react';
-import { i18n } from '../../i18n';
-import { ERainType } from '../interfaces';
+import { ERainType } from './interfaces';
 
 interface HexbinOptionSelectProps {
   callback: (c: ERainType) => void;
