+++ src/vis/sankey/interfaces.ts
@@ -1,6 +1,5 @@
 import { BaseVisConfig, ColumnInfo, ESupportedPlotlyVis } from '../interfaces';
 
-
 export interface ISankeyConfig extends BaseVisConfig {
   type: ESupportedPlotlyVis.SANKEY;
   catColumnsSelected: ColumnInfo[];
