+++ src/vis/stories/Vis/Heatmap/HeatmapRandom.stories.tsx
@@ -2,7 +2,8 @@ import { ComponentStory } from '@storybook/react';
 import * as d3 from 'd3v7';
 import React from 'react';
 import { Vis } from '../../../LazyVis';
-import { BaseVisConfig, EAggregateTypes, EColumnTypes, ENumericalColorScaleType, ESortTypes, ESupportedPlotlyVis, VisColumn } from '../../../interfaces';
+import { ESortTypes } from '../../../heatmap/interfaces';
+import { BaseVisConfig, EAggregateTypes, EColumnTypes, ENumericalColorScaleType, ESupportedPlotlyVis, VisColumn } from '../../../interfaces';
 
 function RNG(seed) {
   const m = 2 ** 35 - 31;
